Restructure code from packages to reduce redundency
