+++ src/main/java/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: thoth.main.Thoth
+Main-Class: thoth.Thoth
