+++ src/main/java/thoth/TaskManager.java
@@ -4,7 +4,7 @@
  * This class provides methods to add tasks, mark them as done or not done, remove tasks,
  * and retrieve tasks or the task count.
  */
-package thoth.logic;
+package thoth;
 
 import thoth.tasks.Task;
 
@@ -61,7 +61,7 @@ public class TaskManager {
     }
 
     /**
-     *  Remove the task at the specific task index
+     * Remove the task at the specific task index
      *
      * @param taskId the index of tht task to be removed
      */
