+++ src/main/java/thoth/Thoth.java
