+++ src/main/java/thoth/exceptions/TaskParsingException.java
@@ -1,0 +1,7 @@
+package thoth.exceptions;
+
+public class TaskParsingException extends RuntimeException {
+  public TaskParsingException(String message) {
+    super(message);
+  }
+}
