+++ src/main/java/thoth/exceptions/ThothException.java
@@ -1,0 +1,7 @@
+package thoth.exceptions;
+
+public class InvalidCommandException extends RuntimeException {
+    public InvalidCommandException(String message) {
+        super(message);
+    }
+}
